Replace request.queryParams type with `any`
The actual type for param's value could be a string, an array or an
object. Also it introduced additional complexity to display this in
desktop app, so the easiest way turned out to be to let server record
query params object as is, and then process it at the app.

This *might* cause some compatibility issues with CLI, but so far test
in CLI make no assumption about how query params are reported
whatsoever.